Move achievement hook to NewBehaviorDialog to reduce duplication
