+++ newIDE/app/src/BehaviorsEditor/index.js
@@ -27,12 +27,6 @@ import EmptyBehaviorsPlaceholder from './EmptyBehaviorsPlaceholder';
 import PreferencesContext from '../MainFrame/Preferences/PreferencesContext';
 import ScrollView from '../UI/ScrollView';
 import { IconContainer } from '../UI/IconContainer';
-import {
-  ACHIEVEMENT_FEATURE_FLAG,
-  addCreateBadgePreHookIfNotClaimed,
-  TRIVIAL_FIRST_BEHAVIOR,
-} from '../Utils/GDevelopServices/Badge';
-import AuthenticatedUserContext from '../Profile/AuthenticatedUserContext';
 import { getBehaviorTutorialIds } from '../Utils/GDevelopServices/Tutorial';
 import { addBehaviorToObject } from '../Utils/Behavior';
 
@@ -52,7 +46,6 @@ const BehaviorsEditor = (props: Props) => {
   const [newBehaviorDialogOpen, setNewBehaviorDialogOpen] = React.useState(
     false
   );
-  const authenticatedUser = React.useContext(AuthenticatedUserContext);
 
   const { object, project } = props;
   const allBehaviorNames = object.getAllBehaviorNames().toJSArray();
@@ -60,7 +53,7 @@ const BehaviorsEditor = (props: Props) => {
 
   const { values } = React.useContext(PreferencesContext);
 
-  const _addBehavior = (type: string, defaultName: string) => {
+  const addBehavior = (type: string, defaultName: string) => {
     setNewBehaviorDialogOpen(false);
 
     addBehaviorToObject(project, object, type, defaultName);
@@ -69,14 +62,6 @@ const BehaviorsEditor = (props: Props) => {
     if (props.onSizeUpdated) props.onSizeUpdated();
     props.onUpdateBehaviorsSharedData();
   };
-
-  const addBehavior = ACHIEVEMENT_FEATURE_FLAG
-    ? addCreateBadgePreHookIfNotClaimed(
-        authenticatedUser,
-        TRIVIAL_FIRST_BEHAVIOR,
-        _addBehavior
-      )
-    : _addBehavior;
 
   const onChangeBehaviorName = (
     behaviorContent: gdBehaviorContent,
